Relocating job scheduler sample extension to plugins/examples/job-scheduler-sample-extension without the test directory as integration/bwd tests are not currently supported for native plugins
