+++ plugins/examples/build.gradle
@@ -36,3 +36,11 @@ configure(project('painless-whitelist')) {
     }
   }
 }
+
+configure(project('job-scheduler-sample-extension')) {
+  configurations.all {
+    resolutionStrategy.dependencySubstitution {
+      substitute module('org.opensearch.plugin:opensearch-job-scheduler-spi') with project(':plugins:job-scheduler:spi')
+    }
+  }
+}
