+++ plugins/examples/job-scheduler-sample-extension/build.gradle
@@ -1,0 +1,27 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+apply plugin: 'opensearch.opensearchplugin'
+apply plugin: 'opensearch.testclusters'
+apply plugin: 'opensearch.java-rest-test'
+
+opensearchplugin {
+  name 'opensearch-job-scheduler-sample-extension'
+  description 'Sample plugin that extends OpenSearch Job Scheduler plugin'
+  classname 'org.opensearch.jobscheduler.sampleextension.SampleExtensionPlugin'
+  extendedPlugins = ['opensearch-job-scheduler']
+  licenseFile rootProject.file('licenses/APACHE-LICENSE-2.0.txt')
+  noticeFile rootProject.file('NOTICE.txt')
+}
+
+dependencies {
+  compileOnly "org.opensearch.plugin:opensearch-job-scheduler-spi:${versions.opensearch}"
+}
+
+testClusters.all {
+  testDistribution = 'archive'
+}
+
+test.enabled = false
