+++ plugins/examples/job-scheduler-sample-extension/src/main/resources/META-INF/services/org.opensearch.jobscheduler.spi.JobSchedulerExtension
@@ -1,0 +1,9 @@
+#
+# SPDX-License-Identifier: Apache-2.0
+# 
+# The OpenSearch Contributors require contributions made to
+# this file be licensed under the Apache-2.0 license or a
+# compatible open source license.
+#
+
+org.opensearch.example.jobschedulersampleextension.SampleExtensionPlugin
