--- plugins/job-scheduler/sample-extension-plugin/src/main/resources/META-INF/services/org.opensearch.jobscheduler.spi.JobSchedulerExtension
@@ -1,6 +1,0 @@
-#
-# Copyright OpenSearch Contributors
-# SPDX-License-Identifier: Apache-2.0
-#
-
-org.opensearch.jobscheduler.sampleextension.SampleExtensionPlugin
