--- plugins/job-scheduler/sample-extension-plugin/src/test/java/org/opensearch/jobscheduler/sampleextension/bwc/JobSchedulerBackwardsCompatibilityIT.java
@@ -1,147 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.opensearch.jobscheduler.sampleextension.bwc;
-
-import org.junit.Assert;
-import org.opensearch.jobscheduler.sampleextension.SampleExtensionIntegTestCase;
-import org.opensearch.test.rest.OpenSearchRestTestCase;
-
-import java.time.Instant;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-
-public class JobSchedulerBackwardsCompatibilityIT extends SampleExtensionIntegTestCase {
-    private static final ClusterType CLUSTER_TYPE = ClusterType.parse(System.getProperty("tests.rest.bwcsuite"));
-    private static final String CLUSTER_NAME = System.getProperty("tests.clustername");
-
-    /*
-    * In this backward compatibility test, only the old version of job-scheduler plugin is loaded in old cluster while both the latest job-scheduler
-    * & sample-extension plugin is loaded in fully upgraded cluster.
-     */
-    @SuppressWarnings("unchecked")
-    public void testBackwardsCompatibility() throws Exception {
-        String uri = getPluginUri();
-        assert uri != null;
-
-        Map<String, Object> response = getAsMap(uri);
-        Map<String, Map<String, Object>> responseMap = (Map<String, Map<String, Object>>) response.get("nodes");
-
-        for (Map<String, Object> respValues: responseMap.values()) {
-            List<Map<String, Object>> plugins = (List<Map<String, Object>>) respValues.get("plugins");
-            List<String> pluginNames = plugins.stream().map(plugin -> plugin.get("name").toString()).collect(Collectors.toList());
-
-            switch (CLUSTER_TYPE) {
-                case OLD:
-                    /*
-                    * as only the old version of job-scheduler plugin is loaded, we only assert that it is loaded.
-                     */
-                    Assert.assertTrue(pluginNames.contains("opendistro-job-scheduler"));
-                    break;
-                case MIXED:
-                    /*
-                    * for a 3-node mixedClusterTask, 1st node is upgraded, & 2 remaining nodes still have old versions.
-                    * for a 3-node twoThirdsUpgradedClusterTask, 2 nodes are upgraded, & 1 remaining node still have old versions.
-                    * for a 3-node rollingUpgradeClusterTask, all 3 nodes are upgraded.
-                    * so for first 2 scenarios, we assert, 3rd node still has older version of job-scheduler plugin & 1st/2nd node is upgraded to use latest plugins.
-                    * we cannot trigger a call for scheduling watcher job, as the older nodes do not have sample-extension plugin.
-                     */
-                    Map<String, Object> responseForOldNode = getAsMap(getPluginUriForMixedCluster("third"));
-                    Map<String, Map<String, Object>> responseMapForOldNode = (Map<String, Map<String, Object>>) responseForOldNode.get("nodes");
-
-                    for (Map<String, Object> respValuesForOldNode: responseMapForOldNode.values()) {
-                        List<Map<String, Object>> pluginsForOldNode = (List<Map<String, Object>>) respValuesForOldNode.get("plugins");
-                        List<String> pluginNamesForOldNode = pluginsForOldNode.stream().map(plugin -> plugin.get("name").toString()).collect(Collectors.toList());
-                        Assert.assertTrue("third".equals(System.getProperty("tests.rest.bwcsuite_round")) ||
-                                pluginNamesForOldNode.contains("opendistro-job-scheduler"));
-                    }
-                case UPGRADED:
-                    /*
-                    * As cluster is fully upgraded either by full restart or rolling upgrade, we assert, that all nodes are upgraded to use latest plugins.
-                    * we trigger a call for scheduling watcher job now.
-                     */
-                    Assert.assertTrue(pluginNames.contains("opensearch-job-scheduler"));
-                    Assert.assertTrue(pluginNames.contains("opensearch-job-scheduler-sample-extension"));
-                    if (CLUSTER_TYPE == ClusterType.UPGRADED || "third".equals(System.getProperty("tests.rest.bwcsuite_round"))) {
-                        createBasicWatcherJob();
-                    }
-            }
-        }
-    }
-
-    private String getPluginUri() {
-        switch (CLUSTER_TYPE) {
-            case OLD:
-                return "_nodes/" + CLUSTER_NAME + "-0/plugins";
-            case MIXED: {
-                return getPluginUriForMixedCluster(System.getProperty("tests.rest.bwcsuite_round"));
-            }
-            case UPGRADED: return "_nodes/plugins";
-        }
-        return null;
-    }
-
-    private String getPluginUriForMixedCluster(String node) {
-        switch (node) {
-            case "second": return "_nodes/" + CLUSTER_NAME + "-1/plugins";
-            case "third": return "_nodes/" + CLUSTER_NAME + "-2/plugins";
-            default: return "_nodes/" + CLUSTER_NAME + "-0/plugins";
-        }
-    }
-
-    private enum ClusterType {
-        OLD,
-        MIXED,
-        UPGRADED;
-        static ClusterType parse(String value) {
-            switch (value) {
-                case "old_cluster":
-                    return OLD;
-                case "mixed_cluster":
-                    return MIXED;
-                case "upgraded_cluster":
-                    return UPGRADED;
-                default:
-                    throw new AssertionError("Unknown cluster type: $value");
-            }
-        }
-    }
-
-    private void createBasicWatcherJob() throws Exception {
-        String index = createTestIndex();
-        Instant now = Instant.now();
-        /*
-        * we insert the oldest version of job metadata directly into the registered index of sample-extension plugin. this will avoid
-        * calling the serde methods of ScheduledJobParameter class.
-        * Once this doc is inserted, JobSweeper listens for changes in the registered index, tries to deserialize, & then schedule the job.
-        * Thus, failure to schedule the job would mean, backward incompatible changes were made in the serde logic.
-        * & the assert would fail.
-         */
-        String jobParameter =
-                 "{" +
-                    "\"name\":\"sample-job-it\"," +
-                    "\"enabled\":true," +
-                    "\"enabled_time\":" + now.toEpochMilli() + ", " +
-                    "\"last_update_time\":" + now.toEpochMilli() + ", " +
-                    "\"schedule\":{" +
-                        "\"interval\":{" +
-                            "\"start_time\":" + now.toEpochMilli() + "," +
-                            "\"period\":1," +
-                            "\"unit\":\"Minutes\"" +
-                        "}" +
-                    "}," +
-                    "\"index_name_to_watch\":\"" + index + "\"," +
-                    "\"lock_duration_seconds\":120" +
-                 "}";
-
-        // Creates a new watcher job.
-        String jobId = OpenSearchRestTestCase.randomAlphaOfLength(10);
-        createWatcherJobJson(jobId, jobParameter);
-
-        long actualCount = waitAndCountRecords(index, 80000);
-        Assert.assertEquals(1, actualCount);
-    }
-}
